# docs/swagger.yaml
basePath: /api/v2/
definitions:
...
  handler.cexAuthorRequest:
    properties:
      author:
        type: string
      wallet_id:
        type: string
    type: object
  handler.cexWalletIDRequest:
    properties:
      wallet_id:
        type: string
    type: object
...
  model.CexConnectRequest:
    properties:
      api_key:
        description: WalletAddress          string   `json:"wallet_address" example:"copytrade-primary"`
        example: cex_api_key
        type: string
      api_secret:
        example: cex_api_secret
        type: string
      exchange:
        example: binance-th
        type: string
      wallet_name:
        example: CEX Main
        type: string
    type: object
  model.CexUpdateAPIKeyRequest:
    properties:
      api_key:
        example: cex_api_key
        type: string
      api_secret:
        example: cex_api_secret
        type: string
      exchange:
        example: binance-th
        type: string
      wallet_id:
        example: e50b0c09-18c5-4ff0-a832-54473e1b739e
        type: string
    type: object
  model.CexUpdateHoldingPeriodRequest:
    properties:
      holding_period:
        type: integer
      wallet_id:
        type: string
    type: object
  model.CexUpdatePositionSizeRequest:
    properties:
      position_size_percentage:
        example: 0.2
        type: number
      wallet_id:
        example: e50b0c09-18c5-4ff0-a832-54473e1b739e
        type: string
    type: object
  model.CexUpdateSLRequest:
    properties:
      exchange:
        example: binance-th
        type: string
      sl_percentage:
        example: 25
        type: number
      wallet_id:
        example: e50b0c09-18c5-4ff0-a832-54473e1b739e
        type: string
    type: object
  model.CexWalletInfo:
    properties:
      authors:
        items:
          $ref: '#/definitions/model.SubscribeAuthor'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      holding_hour_period:
        type: integer
      hyperliquid_basecode:
        type: boolean
      leverage:
        type: integer
      position_size_percentage:
        type: number
      priority:
        type: integer
      privy_wallet_id:
        type: string
      sl_percentage:
        type: number
      tp_percentage:
        type: number
      updated_at:
        type: string
      wallet_id:
        type: string
      wallet_name:
        type: string
      wallet_type:
        type: string
    type: object
  model.CexWalletTotalValue:
    properties:
      total_value:
        type: number
    type: object
...
  model.SubscribeAuthor:
    properties:
      author_username:
        type: string
      id:
        type: string
    type: object
...
info:
  contact:
    email: admintool@investicstudio.com
    name: API Support
  description: API documentation for Crypto Studio Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Studio Backend API
  version: "1.0"
paths:
...
  /cex/active-wallet:
    post:
      consumes:
      - application/json
      description: Activate the selected CEX wallet for the current user
      parameters:
      - description: CEX wallet id
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.cexWalletIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Activate CEX wallet
      tags:
      - copytrade/cex
  /cex/add-wallet:
    post:
      consumes:
      - application/json
      description: Link a CEX account to the current authenticated user
      parameters:
      - description: CEX connect payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CexConnectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Connect CEX wallet
      tags:
      - copytrade/cex
  /cex/deactive-wallet:
    post:
      consumes:
      - application/json
      description: Deactivate the selected CEX wallet for the current user
      parameters:
      - description: CEX wallet id
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.cexWalletIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate CEX wallet
      tags:
      - copytrade/cex
  /cex/subscribe-author:
    post:
      consumes:
      - application/json
      description: Subscribe an author to a CEX wallet
      parameters:
      - description: Subscribe author payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.cexAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe author
      tags:
      - copytrade/cex
  /cex/unsubscribe-author:
    post:
      consumes:
      - application/json
      description: Unsubscribe an author from a CEX wallet
      parameters:
      - description: Unsubscribe author payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.cexAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe author
      tags:
      - copytrade/cex
  /cex/update-api-key:
    post:
      consumes:
      - application/json
      description: Update the API key and secret of a CEX wallet
      parameters:
      - description: API credential payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CexUpdateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update API credentials
      tags:
      - copytrade/cex
  /cex/update-holding-period:
    post:
      consumes:
      - application/json
      description: Update the holding period (in hours) configured for a CEX wallet
      parameters:
      - description: Holding period payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CexUpdateHoldingPeriodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update holding period
      tags:
      - copytrade/cex
  /cex/update-position-size:
    post:
      consumes:
      - application/json
      description: Update the position size percentage of a CEX wallet
      parameters:
      - description: Position size payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CexUpdatePositionSizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update position size percentage
      tags:
      - copytrade/cex
  /cex/update-sl:
    post:
      consumes:
      - application/json
      description: Update the stop loss percentage of a CEX wallet
      parameters:
      - description: Stop loss payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CexUpdateSLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update stop loss
      tags:
      - copytrade/cex
  /cex/wallet-info:
    get:
      description: Retrieve CEX wallets for the current authenticated user filtered
        by exchange
      parameters:
      - description: Exchange identifier (e.g. bybit)
        in: query
        name: exchange
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CexWalletInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List CEX wallets
      tags:
      - copytrade/cex
  /cex/wallet-total-value:
    get:
      description: Retrieve the total value of a CEX wallet for the current authenticated
        user
      parameters:
      - description: CEX wallet ID
        in: query
        name: wallet_id
        required: true
        type: string
      - description: Exchange identifier (e.g. binance-th)
        in: query
        name: exchange
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CexWalletTotalValue'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get wallet total value
      tags:
      - copytrade/cex
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your-token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"